+++ contracts/test/PBHExternalNullifier.t.sol
@@ -6,11 +6,11 @@ import "@lib/PBHExternalNullifier.sol";
 import "@BokkyPooBahsDateTimeLibrary/BokkyPooBahsDateTimeLibrary.sol";
 
 contract CallDepth1 {
-    function encodeRevertCallDepth1(uint8 pbhNonce, uint8 month, uint16 year) public pure {
+    function encodeRevertCallDepth1(uint16 pbhNonce, uint8 month, uint16 year) public pure {
         PBHExternalNullifier.encode(PBHExternalNullifier.V1, pbhNonce, month, year);
     }
 
-    function verifyRevertCallDepth1(uint256 encoded, uint8 maxPbh, uint256 signalHash) public view {
+    function verifyRevertCallDepth1(uint256 encoded, uint16 maxPbh, uint256 signalHash) public view {
         PBHExternalNullifier.verify(encoded, maxPbh, signalHash);
     }
 }
@@ -19,23 +19,23 @@ contract CallDepth1 {
 /// @notice Contains tests for the PBHExternalNullifier library
 /// @author Worldcoin
 contract PBHExternalNullifierTest is Test {
-    function testFuzz_encode(uint8 pbhNonce, uint8 month, uint16 year) public pure {
+    function testFuzz_encode(uint16 pbhNonce, uint8 month, uint16 year) public pure {
         vm.assume(month > 0 && month <= 12);
         PBHExternalNullifier.encode(PBHExternalNullifier.V1, pbhNonce, month, year);
     }
 
-    function testFuzz_encode_RevertIf_InvalidMonth(uint8 pbhNonce, uint8 month, uint16 year) public {
+    function testFuzz_encode_RevertIf_InvalidMonth(uint16 pbhNonce, uint8 month, uint16 year) public {
         vm.assume(month == 0 || month > 12);
         CallDepth1 callDepth1 = new CallDepth1();
         vm.expectRevert(PBHExternalNullifier.InvalidExternalNullifierMonth.selector);
         callDepth1.encodeRevertCallDepth1(pbhNonce, month, year);
     }
 
-    function testFuzz_decode(uint8 pbhNonce, uint8 month, uint16 year) public {
+    function testFuzz_decode(uint16 pbhNonce, uint8 month, uint16 year) public {
         vm.assume(month > 0 && month <= 12);
         uint256 encoded = PBHExternalNullifier.encode(PBHExternalNullifier.V1, pbhNonce, month, year);
 
-        (uint8 decodedVersion, uint8 decodedNonce, uint8 decodedMonth, uint16 decodedYear) =
+        (uint8 decodedVersion, uint16 decodedNonce, uint8 decodedMonth, uint16 decodedYear) =
             PBHExternalNullifier.decode(encoded);
 
         assertEq(decodedVersion, PBHExternalNullifier.V1);
@@ -44,7 +44,7 @@ contract PBHExternalNullifierTest is Test {
         assertEq(decodedYear, year);
     }
 
-    function testFuzz_verify(uint8 pbhNonce, uint8 month, uint16 year, uint8 maxPbh) public {
+    function testFuzz_verify(uint16 pbhNonce, uint8 month, uint16 year, uint16 maxPbh) public {
         vm.assume(month > 0 && month <= 12);
         vm.assume(year >= 2023);
         vm.assume(maxPbh > 0 && pbhNonce < maxPbh);
@@ -58,7 +58,7 @@ contract PBHExternalNullifierTest is Test {
     }
 
     function testFuzz_verify_RevertIf_InvalidNullifierLeadingZeros(uint256 encoded) public {
-        vm.assume(encoded > type(uint40).max);
+        vm.assume(encoded > type(uint48).max);
         CallDepth1 callDepth1 = new CallDepth1();
         vm.expectRevert(
             abi.encodeWithSelector(PBHExternalNullifier.InvalidExternalNullifier.selector, encoded, 0, "Leading zeros")
@@ -72,7 +72,7 @@ contract PBHExternalNullifierTest is Test {
         uint8 month = uint8(BokkyPooBahsDateTimeLibrary.getMonth(block.timestamp));
         uint16 year = uint16(BokkyPooBahsDateTimeLibrary.getYear(block.timestamp));
 
-        uint8 pbhNonce = 0;
+        uint16 pbhNonce = 0;
         uint8 maxPbh = 30;
         uint256 encoded = PBHExternalNullifier.encode(pbhVersion, pbhNonce, month, year);
         CallDepth1 callDepth1 = new CallDepth1();
@@ -92,8 +92,8 @@ contract PBHExternalNullifierTest is Test {
         uint256 timestamp = BokkyPooBahsDateTimeLibrary.timestampFromDate(year + 1, month, 1);
         vm.warp(timestamp);
 
-        uint8 pbhNonce = 0;
-        uint8 maxPbh = 30;
+        uint16 pbhNonce = 0;
+        uint16 maxPbh = 30;
         uint256 encoded = PBHExternalNullifier.encode(PBHExternalNullifier.V1, pbhNonce, month, year);
         CallDepth1 callDepth1 = new CallDepth1();
         vm.expectRevert(
@@ -110,8 +110,8 @@ contract PBHExternalNullifierTest is Test {
         uint256 timestamp = BokkyPooBahsDateTimeLibrary.timestampFromDate(year, month + 1, 1);
         vm.warp(timestamp);
 
-        uint8 pbhNonce = 0;
-        uint8 maxPbh = 30;
+        uint16 pbhNonce = 0;
+        uint16 maxPbh = 30;
         uint256 encoded = PBHExternalNullifier.encode(PBHExternalNullifier.V1, pbhNonce, month, year);
         CallDepth1 callDepth1 = new CallDepth1();
         vm.expectRevert(
@@ -120,7 +120,7 @@ contract PBHExternalNullifierTest is Test {
         callDepth1.verifyRevertCallDepth1(encoded, maxPbh, 0);
     }
 
-    function testFuzz_verify_RevertIf_InvalidPbhNonce(uint8 pbhNonce, uint8 maxPbh) public {
+    function testFuzz_verify_RevertIf_InvalidPbhNonce(uint16 pbhNonce, uint8 maxPbh) public {
         vm.assume(maxPbh > 0 && pbhNonce >= maxPbh);
 
         uint8 month = uint8(BokkyPooBahsDateTimeLibrary.getMonth(block.timestamp));
